==> data/00_dymmy_data.ipynb <==
import pandas as pd

# create a dataframe and save as csv
df = pd.DataFrame({'a': [1, 2, 3], 'b': [4, 5, 6]})
df.to_csv('test.csv')
df.to_csv('test.csv.gzip',compression='gzip',index=False)

# create a dataframe with two date columns and save as csv

df = pd.DataFrame({'a': [1, 2, 3], 'b': [4, 5, 6], 'c': ['2019-01-01', '2019-01-02', '2019-01-03'], 'd': ['2019-01-01', '2019-01-02', '2019-01-03']})
df.to_csv('test_date.csv')

# create a dataframe with string, number and dates and save as csv

df = pd.DataFrame({'a': [1, 2, 3], 'b': [4, 5, 6], 'c': ['2019-01-01', '2019-01-02', '2019-01-03'], 'd': ['2019-01-01', '2019-01-02', '2019-01-03'], 'e': ['a', 'b', 'c']})
df.to_csv('test_date_str.csv', index=False)
